+++ src/main/java/com/getindata/connectors/http/HttpSink.java
@@ -40,7 +40,7 @@ public class HttpSink<InputT> extends HttpSinkInternal<InputT> {
             long maxTimeInBufferMS,
             long maxRecordSizeInBytes,
             String endpointUrl,
-            HttpSinkPostRequestCallback httpSinkPostRequestCallback,
+            HttpPostRequestCallback httpPostRequestCallback,
             SinkHttpClientBuilder sinkHttpClientBuilder,
             Properties properties) {
         super(elementConverter,
@@ -51,7 +51,7 @@ public class HttpSink<InputT> extends HttpSinkInternal<InputT> {
             maxTimeInBufferMS,
             maxRecordSizeInBytes,
             endpointUrl,
-              httpSinkPostRequestCallback,
+            httpPostRequestCallback,
             sinkHttpClientBuilder,
             properties
         );