+++ src/main/java/com/getindata/connectors/http/HttpSinkBuilder.java
@@ -70,7 +70,7 @@ public class HttpSinkBuilder<InputT> extends
     // If not defined, should be set to DEFAULT_CLIENT_BUILDER
     private SinkHttpClientBuilder sinkHttpClientBuilder;
 
-    private HttpSinkPostRequestCallback httpSinkPostRequestCallback;
+    private HttpPostRequestCallback httpPostRequestCallback;
 
     HttpSinkBuilder() {
         this.sinkHttpClientBuilder = DEFAULT_CLIENT_BUILDER;
@@ -107,10 +107,9 @@ public class HttpSinkBuilder<InputT> extends
         return this;
     }
 
-    public HttpSinkBuilder<InputT> setHttpSinkPostRequestCallback(
-        HttpSinkPostRequestCallback httpSinkPostRequestCallback
-    ) {
-        this.httpSinkPostRequestCallback = httpSinkPostRequestCallback;
+    public HttpSinkBuilder<InputT> setHttpPostRequestCallback(
+        HttpPostRequestCallback httpPostRequestCallback) {
+        this.httpPostRequestCallback = httpPostRequestCallback;
         return this;
     }
 
@@ -146,7 +145,7 @@ public class HttpSinkBuilder<InputT> extends
             Optional.ofNullable(getMaxTimeInBufferMS()).orElse(DEFAULT_MAX_TIME_IN_BUFFER_MS),
             Optional.ofNullable(getMaxRecordSizeInBytes()).orElse(DEFAULT_MAX_RECORD_SIZE_IN_B),
             endpointUrl,
-            httpSinkPostRequestCallback,
+            httpPostRequestCallback,
             sinkHttpClientBuilder,
             properties
         );