+++ src/main/java/com/getindata/connectors/http/internal/table/sink/Slf4jHttpPostRequestCallback.java
@@ -6,18 +6,24 @@ import java.util.Map;
 
 import lombok.extern.slf4j.Slf4j;
 
-import com.getindata.connectors.http.HttpSinkPostRequestCallback;
+import com.getindata.connectors.http.HttpPostRequestCallback;
 import com.getindata.connectors.http.internal.sink.HttpSinkRequestEntry;
 
+/**
+ * A {@link HttpPostRequestCallback} that logs pairs of request and response as <i>INFO</i> level
+ * logs using <i>Slf4j</i>.
+ *
+ * <p>Serving as a default implementation of {@link HttpPostRequestCallback} for
+ * the {@link HttpDynamicSink}.
+ */
 @Slf4j
-public class Slf4jHttpSinkPostRequestCallback implements HttpSinkPostRequestCallback {
+public class Slf4jHttpPostRequestCallback implements HttpPostRequestCallback {
     @Override
     public void call(
         HttpResponse<String> response,
         HttpSinkRequestEntry requestEntry,
         String _endpointUrl,
-        Map<String, String> _headerMap
-    ) {
+        Map<String, String> _headerMap) {
         log.info(
             "Got response for a request.\n  Request:\n    " +
                 "Method: {}\n    Body: {}\n  Response: {}\n    Body: {}",