+++ src/test/java/com/getindata/connectors/http/internal/sink/HttpSinkBuilderTest.java
@@ -9,8 +9,8 @@ import org.apache.flink.connector.base.sink.writer.ElementConverter;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import com.getindata.connectors.http.HttpPostRequestCallback;
 import com.getindata.connectors.http.HttpSink;
-import com.getindata.connectors.http.HttpSinkPostRequestCallback;
 import com.getindata.connectors.http.internal.SinkHttpClient;
 import com.getindata.connectors.http.internal.SinkHttpClientResponse;
 
@@ -57,7 +57,8 @@ public class HttpSinkBuilderTest {
     private static class MockHttpClient implements SinkHttpClient {
 
         MockHttpClient(
-            Properties properties, HttpSinkPostRequestCallback httpSinkPostRequestCallback) {
+            Properties properties, HttpPostRequestCallback httpPostRequestCallback
+        ) {
 
         }