+++ src/test/java/com/getindata/connectors/http/internal/table/sink/HttpDynamicSinkTest.java
@@ -70,7 +70,7 @@ public class HttpDynamicSinkTest {
             )
             .setConsumedDataType(consumedDataType)
             .setEncodingFormat(mockFormat)
-            .setHttpSinkPostRequestCallback(new Slf4jHttpSinkPostRequestCallback())
+            .setHttpPostRequestCallback(new Slf4jHttpPostRequestCallback())
             .setMaxBatchSize(1);
     }
 
@@ -91,9 +91,9 @@ public class HttpDynamicSinkTest {
             getSinkBuilder().setConsumedDataType(new AtomicDataType(new BooleanType(true))).build();
         var sinkFormat = getSinkBuilder().setEncodingFormat(
             new TestFormatFactory.EncodingFormatMock(";", ChangelogMode.all())).build();
-        var sinkHttpSinkPostRequestCallback =
+        var sinkHttpPostRequestCallback =
             getSinkBuilder()
-                .setHttpSinkPostRequestCallback(new Slf4jHttpSinkPostRequestCallback()).build();
+                .setHttpPostRequestCallback(new Slf4jHttpPostRequestCallback()).build();
 
         assertEquals(sink, sink);
         assertNotEquals(null, sink);
@@ -102,6 +102,6 @@ public class HttpDynamicSinkTest {
         assertNotEquals(sink, sinkSinkConfig);
         assertNotEquals(sink, sinkDataType);
         assertNotEquals(sink, sinkFormat);
-        assertNotEquals(sink, sinkHttpSinkPostRequestCallback);
+        assertNotEquals(sink, sinkHttpPostRequestCallback);
     }
 }